Add tests for outside click and constants variables
