+++ dev/tests/contants.js
@@ -1,0 +1,18 @@
+export const options = [
+  {
+    name: 'option 1',
+  },
+  {
+    name: 'option 2',
+  },
+  {
+    name: 'option 3',
+  },
+];
+
+export const [selectedOption] = options;
+
+export default {
+  options,
+  selectedOption,
+};
