+++ dev/tests/unit/dropdown-render.spec.js
@@ -1,21 +1,8 @@
 import { mount } from '@vue/test-utils';
 import VueDropdown from '../../../Dropdown.vue';
+import {options, selectedOption} from '../contants';
 
-const options = [
-  {
-    name: 'option 1',
-  },
-  {
-    name: 'option 2',
-  },
-  {
-    name: 'option 3',
-  },
-];
-
-const [selectedOption] = options;
-
-describe('Dropdown.vue', () => {
+describe('Dropdown.vue render', () => {
   it('renders dropdown-toggle', () => {
     const wrapper = mount(VueDropdown, {
       propsData: {
@@ -61,29 +48,5 @@ describe('Dropdown.vue', () => {
     });
 
     expect(allItemsRendered).toBe(true);
-  });
-
-  it('fires updateOption event on click event item', async () => {
-    const updateOptionHandler = jest.fn();
-
-    const wrapper = mount(VueDropdown, {
-      propsData: {
-        options,
-        selected: selectedOption,
-      },
-      listeners: {
-        updateOption: updateOptionHandler,
-      }
-    });
-    const dropdownToggle = wrapper.find('.dropdown-toggle');
-    dropdownToggle.trigger('click');
-
-    await wrapper.vm.$nextTick();
-
-    const optionItems = wrapper.findAll('.dropdown-menu li a');
-    const secondItem = optionItems.at(1);
-    secondItem.trigger('click');
-
-    expect(updateOptionHandler.mock.calls.length).toBe(1);
   });
 });
